web/requirements.txt
firebase-admin
google-cloud-firestore
Flask==1.0.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
Flask-MySQLdb==0.2.0
passlib==1.7.1
requests==2.20.0
pillow
gunicorn
